dom/bindings/test/test_bug1123875.html

<!doctype html>
<meta charset=utf-8>
<title>Test for Bug 1123875</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
  test(() => {
    assert_throws(new TypeError, () => {
      "use strict";
      document.childNodes.length = 0;
    });
  }, "setting a readonly attribute on a proxy in strict mode should throw a TypeError");
</script>
